Move timed-out.ts out of utils
