+++ source/core/timed-out.ts
@@ -1,6 +1,6 @@
 import * as net from 'net';
 import {ClientRequest, IncomingMessage} from 'http';
-import unhandler from './unhandle';
+import unhandler from './utils/unhandle';
 
 const reentry: unique symbol = Symbol('reentry');
 const noop = (): void => {};
